Replace barrier with cpp20_barrier
