+++ libs/pika/algorithms/include/pika/parallel/spmd_block.hpp
@@ -40,7 +40,7 @@ namespace pika {
     struct spmd_block
     {
     private:
-        using barrier_type = pika::lcos::barrier;
+        using barrier_type = pika::barrier<>;
         using table_type =
             std::map<std::set<std::size_t>, std::shared_ptr<barrier_type>>;
         using mutex_type = pika::mutex;
@@ -77,7 +77,7 @@ namespace pika {
 
         void sync_all() const
         {
-            barrier_.get().wait();
+            barrier_.get().arrive_and_wait();
         }
 
         void sync_images(std::set<std::size_t> const& images) const
@@ -103,7 +103,7 @@ namespace pika {
 
             if (images.find(image_id_) != images.end())
             {
-                it->second->wait();
+                it->second->arrive_and_wait();
             }
         }
 
@@ -145,7 +145,7 @@ namespace pika {
         struct spmd_block_helper
         {
         private:
-            using barrier_type = pika::lcos::barrier;
+            using barrier_type = pika::barrier<>;
             using table_type =
                 std::map<std::set<std::size_t>, std::shared_ptr<barrier_type>>;
             using mutex_type = pika::mutex;
@@ -180,7 +180,7 @@ namespace pika {
             using first_type =
                 typename pika::util::detail::first_argument<ftype>::type;
 
-            using barrier_type = pika::lcos::barrier;
+            using barrier_type = pika::barrier<>;
             using table_type =
                 std::map<std::set<std::size_t>, std::shared_ptr<barrier_type>>;
             using mutex_type = pika::mutex;
@@ -217,7 +217,7 @@ namespace pika {
             using first_type =
                 typename pika::util::detail::first_argument<ftype>::type;
 
-            using barrier_type = pika::lcos::barrier;
+            using barrier_type = pika::barrier<>;
             using table_type =
                 std::map<std::set<std::size_t>, std::shared_ptr<barrier_type>>;
             using mutex_type = pika::mutex;
