+++ libs/pika/synchronization/include/pika/synchronization/barrier.hpp
@@ -23,7 +23,7 @@
 #include <pika/config/warnings_prefix.hpp>
 
 ///////////////////////////////////////////////////////////////////////////////
-namespace pika::lcos {
+namespace pika {
 
     namespace detail {
 
@@ -89,10 +89,10 @@ namespace pika::lcos {
     // Cpp17Destructible (Table 32) requirements.
 
     template <typename OnCompletion = detail::empty_oncompletion>
-    class cpp20_barrier
+    class barrier
     {
     public:
-        PIKA_NON_COPYABLE(cpp20_barrier);
+        PIKA_NON_COPYABLE(barrier);
 
     private:
         using mutex_type = pika::spinlock;
@@ -116,7 +116,7 @@ namespace pika::lcos {
         //                 end note]
         // Throws:         Any exception thrown by CompletionFunction's move
         //                 constructor.
-        cpp20_barrier(
+        barrier(
             std::ptrdiff_t expected, OnCompletion completion = OnCompletion())
           : expected_(expected)
           , arrived_(expected)
@@ -187,7 +187,7 @@ namespace pika::lcos {
             std::unique_lock<mutex_type> l(mtx_);
             if (phase_ == old_phase)
             {
-                cond_.wait(l, "cpp20_barrier::wait");
+                cond_.wait(l, "barrier::wait");
                 PIKA_ASSERT(phase_ != old_phase);
             }
         }
@@ -230,57 +230,6 @@ namespace pika::lcos {
         bool phase_;
     };
 
-    ///////////////////////////////////////////////////////////////////////////
-    /// A barrier can be used to synchronize a specific number of threads,
-    /// blocking all of the entering threads until all of the threads have
-    /// entered the barrier.
-    ///
-    /// \note   A \a barrier is not a LCO in the sense that it has no global id
-    ///         and it can't be triggered using the action (parcel) mechanism.
-    ///         It is just a low level synchronization primitive allowing to
-    ///         synchronize a given number of \a threads.
-    class PIKA_EXPORT barrier
-    {
-    private:
-        using mutex_type = pika::spinlock;
-
-        static constexpr std::size_t barrier_flag = static_cast<std::size_t>(1)
-            << (CHAR_BIT * sizeof(std::size_t) - 1);
-
-    public:
-        barrier(std::size_t number_of_threads);
-        ~barrier();
-
-        /// The function \a wait will block the number of entering \a threads
-        /// (as given by the constructor parameter \a number_of_threads),
-        /// releasing all waiting threads as soon as the last \a thread
-        /// entered this function.
-        void wait();
-
-        /// The function \a count_up will increase the number of \a threads
-        /// to be waited in \a wait function.
-        void count_up();
-
-        /// The function \a reset will reset the number of \a threads
-        /// as given by the function parameter \a number_of_threads.
-        /// the newer coming \a threads executing the function
-        /// \a wait will be waiting until \a total_ is equal to  \a barrier_flag.
-        /// The last \a thread exiting the \a wait function will notify
-        /// the newer \a threads waiting and the newer \a threads
-        /// will get the reset \a number_of_threads_.
-        /// The function \a reset can be executed while previous \a threads
-        /// executing waiting after they have been waken up.
-        /// Thus \a total_ can not be reset to \a barrier_flag which
-        /// will break the comparison condition under the function \a wait.
-        void reset(std::size_t number_of_threads);
-
-    private:
-        std::size_t number_of_threads_;
-        std::size_t total_;
-
-        mutable mutex_type mtx_;
-        pika::detail::condition_variable cond_;
-    };
-}    // namespace pika::lcos
+}    // namespace pika
 
 #include <pika/config/warnings_suffix.hpp>
