+++ libs/pika/synchronization/tests/unit/barrier.cpp
@@ -1,104 +1,219 @@
-//  Copyright (c) 2007-2012 Hartmut Kaiser
-//  Copyright (c) 2011-2012 Bryce Adelstein-Lelbach
+//  Copyright (c) 2020 Hartmut Kaiser
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #include <pika/barrier.hpp>
-#include <pika/functional/bind.hpp>
 #include <pika/init.hpp>
+#include <pika/modules/async.hpp>
 #include <pika/testing.hpp>
-#include <pika/thread.hpp>
 
 #include <atomic>
 #include <cstddef>
 #include <functional>
 #include <string>
+#include <utility>
 #include <vector>
 
-using pika::program_options::options_description;
-using pika::program_options::value;
-using pika::program_options::variables_map;
-
-using pika::threads::detail::make_thread_function_nullary;
-using pika::threads::detail::register_work;
-using pika::threads::detail::thread_init_data;
-
-using pika::lcos::barrier;
-
-using pika::finalize;
-using pika::init;
+std::atomic<std::size_t> c1(0);
+std::atomic<std::size_t> c2(0);
 
 ///////////////////////////////////////////////////////////////////////////////
-void local_barrier_test(barrier& b, std::atomic<std::size_t>& c)
+void local_barrier_test_no_completion(pika::barrier<>& b)
 {
-    ++c;
+    ++c1;
+
     // wait for all threads to enter the barrier
-    b.wait();
+    b.arrive_and_wait();
+
+    ++c2;
 }
 
-///////////////////////////////////////////////////////////////////////////////
-int pika_main(variables_map& vm)
+void test_barrier_empty_oncomplete()
 {
-    std::size_t pxthreads = 0;
+    constexpr std::size_t threads = 64;
+    constexpr std::size_t iterations = 100;
 
-    if (vm.count("pxthreads"))
-        pxthreads = vm["pxthreads"].as<std::size_t>();
-
-    std::size_t iterations = 0;
-
-    if (vm.count("iterations"))
-        iterations = vm["iterations"].as<std::size_t>();
-
-    for (std::size_t i = 0; i < iterations; ++i)
+    for (std::size_t i = 0; i != iterations; ++i)
     {
         // create a barrier waiting on 'count' threads
-        barrier b(pxthreads + 1);
-
-        std::atomic<std::size_t> c(0);
+        pika::barrier<> b(threads + 1);
+        c1 = 0;
+        c2 = 0;
 
         // create the threads which will wait on the barrier
-        for (std::size_t i = 0; i < pxthreads; ++i)
+        std::vector<pika::future<void>> results;
+        results.reserve(threads);
+        for (std::size_t i = 0; i != threads; ++i)
         {
-            thread_init_data data(
-                make_thread_function_nullary(pika::util::detail::bind(
-                    &local_barrier_test, std::ref(b), std::ref(c))),
-                "local_barrier_test");
-            register_work(data);
+            results.push_back(
+                pika::async(&local_barrier_test_no_completion, std::ref(b)));
         }
 
-        b.wait();    // wait for all threads to enter the barrier
-        PIKA_TEST_EQ(pxthreads, c);
-    }
+        b.arrive_and_wait();    // wait for all threads to enter the barrier
+        PIKA_TEST_EQ(threads, c1);
 
-    // initiate shutdown of the runtime system
-    finalize();
-    return 0;
+        pika::wait_all(results);
+
+        PIKA_TEST_EQ(threads, c2);
+    }
 }
 
 ///////////////////////////////////////////////////////////////////////////////
+std::atomic<std::size_t> complete(0);
+
+struct oncomplete
+{
+    void operator()() const
+    {
+        ++complete;
+    }
+};
+
+void local_barrier_test(pika::barrier<oncomplete>& b)
+{
+    ++c1;
+
+    // wait for all threads to enter the barrier
+    b.arrive_and_wait();
+
+    ++c2;
+}
+
+void test_barrier_oncomplete()
+{
+    constexpr std::size_t threads = 64;
+    constexpr std::size_t iterations = 100;
+
+    for (std::size_t i = 0; i != iterations; ++i)
+    {
+        // create a barrier waiting on 'count' threads
+        pika::barrier<oncomplete> b(threads + 1);
+        c1 = 0;
+        c2 = 0;
+        complete = 0;
+
+        // create the threads which will wait on the barrier
+        std::vector<pika::future<void>> results;
+        results.reserve(threads);
+        for (std::size_t i = 0; i != threads; ++i)
+        {
+            results.push_back(pika::async(&local_barrier_test, std::ref(b)));
+        }
+
+        b.arrive_and_wait();    // wait for all threads to enter the barrier
+        PIKA_TEST_EQ(threads, c1);
+
+        pika::wait_all(results);
+
+        PIKA_TEST_EQ(threads, c2);
+        PIKA_TEST_EQ(complete, std::size_t(1));
+    }
+}
+
+///////////////////////////////////////////////////////////////////////////////
+void local_barrier_test_no_completion_split(pika::barrier<>& b)
+{
+    ++c1;
+
+    // signal the barrier
+    auto token = b.arrive();
+
+    // wait for all threads to enter the barrier
+    b.wait(std::move(token));
+
+    ++c2;
+}
+
+void test_barrier_empty_oncomplete_split()
+{
+    constexpr std::size_t threads = 64;
+    constexpr std::size_t iterations = 100;
+
+    for (std::size_t i = 0; i != iterations; ++i)
+    {
+        // create a barrier waiting on 'count' threads
+        pika::barrier<> b(threads + 1);
+        c1 = 0;
+        c2 = 0;
+
+        // create the threads which will wait on the barrier
+        std::vector<pika::future<void>> results;
+        results.reserve(threads);
+        for (std::size_t i = 0; i != threads; ++i)
+        {
+            results.push_back(pika::async(
+                &local_barrier_test_no_completion_split, std::ref(b)));
+        }
+
+        b.arrive_and_wait();    // wait for all threads to enter the barrier
+        PIKA_TEST_EQ(threads, c1);
+
+        pika::wait_all(results);
+
+        PIKA_TEST_EQ(threads, c2);
+    }
+}
+
+void local_barrier_test_split(pika::barrier<oncomplete>& b)
+{
+    ++c1;
+
+    // signal the barrier
+    auto token = b.arrive();
+
+    // wait for all threads to enter the barrier
+    b.wait(std::move(token));
+
+    ++c2;
+}
+
+void test_barrier_oncomplete_split()
+{
+    constexpr std::size_t threads = 64;
+    constexpr std::size_t iterations = 100;
+
+    for (std::size_t i = 0; i != iterations; ++i)
+    {
+        // create a barrier waiting on 'count' threads
+        pika::barrier<oncomplete> b(threads + 1);
+        c1 = 0;
+        c2 = 0;
+        complete = 0;
+
+        // create the threads which will wait on the barrier
+        std::vector<pika::future<void>> results;
+        results.reserve(threads);
+        for (std::size_t i = 0; i != threads; ++i)
+        {
+            results.push_back(
+                pika::async(&local_barrier_test_split, std::ref(b)));
+        }
+
+        b.arrive_and_wait();    // wait for all threads to enter the barrier
+        PIKA_TEST_EQ(threads, c1);
+
+        pika::wait_all(results);
+
+        PIKA_TEST_EQ(threads, c2);
+        PIKA_TEST_EQ(complete, std::size_t(1));
+    }
+}
+
+///////////////////////////////////////////////////////////////////////////////
+int pika_main()
+{
+    test_barrier_empty_oncomplete();
+    test_barrier_oncomplete();
+
+    test_barrier_empty_oncomplete_split();
+    test_barrier_oncomplete_split();
+
+    return pika::finalize();
+}
+
 int main(int argc, char* argv[])
 {
-    // Configure application-specific options
-    options_description desc_commandline(
-        "Usage: " PIKA_APPLICATION_STRING " [options]");
-
-    desc_commandline.add_options()("pxthreads,T",
-        value<std::size_t>()->default_value(64),
-        "the number of PX threads to invoke")("iterations",
-        value<std::size_t>()->default_value(64),
-        "the number of times to repeat the test");
-
-    // We force this test to use several threads by default.
-    std::vector<std::string> const cfg = {"pika.os_threads=all"};
-
-    // Initialize and run pika
-    pika::init_params init_args;
-    init_args.desc_cmdline = desc_commandline;
-    init_args.cfg = cfg;
-
-    PIKA_TEST_EQ_MSG(pika::init(pika_main, argc, argv, init_args), 0,
-        "pika main exited with non-zero status");
-    return 0;
+    return pika::init(pika_main, argc, argv);
 }
