--- libs/pika/synchronization/tests/unit/barrier_count_up.cpp
@@ -1,109 +1,0 @@
-//  Copyright (c) 2007-2012 Hartmut Kaiser
-//  Copyright (c) 2011-2012 Bryce Adelstein-Lelbach
-//  Copyright (c) 2019 Tianyi Zhang
-//
-//  SPDX-License-Identifier: BSL-1.0
-//  Distributed under the Boost Software License, Version 1.0. (See accompanying
-//  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-
-#include <pika/barrier.hpp>
-#include <pika/functional/bind.hpp>
-#include <pika/future.hpp>
-#include <pika/init.hpp>
-#include <pika/testing.hpp>
-#include <pika/thread.hpp>
-
-#include <atomic>
-#include <cstddef>
-#include <functional>
-#include <string>
-#include <vector>
-
-using pika::program_options::options_description;
-using pika::program_options::value;
-using pika::program_options::variables_map;
-
-using pika::threads::detail::make_thread_function_nullary;
-using pika::threads::detail::register_work;
-using pika::threads::detail::thread_init_data;
-
-using pika::lcos::barrier;
-
-using pika::finalize;
-using pika::init;
-
-///////////////////////////////////////////////////////////////////////////////
-void local_barrier_test(barrier& b, std::atomic<std::size_t>& c)
-{
-    ++c;
-    // wait for all threads to enter the barrier
-    b.wait();
-}
-
-///////////////////////////////////////////////////////////////////////////////
-int pika_main(variables_map& vm)
-{
-    std::size_t pxthreads = 0;
-
-    if (vm.count("pxthreads"))
-        pxthreads = vm["pxthreads"].as<std::size_t>();
-
-    std::size_t iterations = 0;
-
-    if (vm.count("iterations"))
-        iterations = vm["iterations"].as<std::size_t>();
-
-    for (std::size_t i = 0; i < iterations; ++i)
-    {
-        // create a barrier waiting on 'this' thread
-        barrier b(1);
-
-        std::atomic<std::size_t> c(0);
-
-        // create the threads which will wait on the barrier
-        for (std::size_t i = 0; i < pxthreads; ++i)
-        {
-            //call count_up to increase number of threads waiting when create a new thread
-            b.count_up();
-
-            thread_init_data data(
-                make_thread_function_nullary(pika::util::detail::bind(
-                    &local_barrier_test, std::ref(b), std::ref(c))),
-                "local_barrier_test");
-            register_work(data);
-        }
-
-        b.wait();    // wait for all threads to enter the barrier
-        PIKA_TEST_EQ(pxthreads, c);
-    }
-
-    // initiate shutdown of the runtime system
-    finalize();
-    return 0;
-}
-
-///////////////////////////////////////////////////////////////////////////////
-int main(int argc, char* argv[])
-{
-    // Configure application-specific options
-    options_description desc_commandline(
-        "Usage: " PIKA_APPLICATION_STRING " [options]");
-
-    desc_commandline.add_options()("pxthreads,T",
-        value<std::size_t>()->default_value(64),
-        "the number of PX threads to invoke")("iterations",
-        value<std::size_t>()->default_value(64),
-        "the number of times to repeat the test");
-
-    // We force this test to use several threads by default.
-    std::vector<std::string> const cfg = {"pika.os_threads=all"};
-
-    // Initialize and run pika
-    pika::init_params init_args;
-    init_args.desc_cmdline = desc_commandline;
-    init_args.cfg = cfg;
-
-    PIKA_TEST_EQ_MSG(pika::init(pika_main, argc, argv, init_args), 0,
-        "pika main exited with non-zero status");
-    return 0;
-}
