+++ libs/pika/threading/tests/unit/thread_id.cpp
@@ -19,12 +19,12 @@ using pika::program_options::options_description;
 using pika::program_options::variables_map;
 
 ///////////////////////////////////////////////////////////////////////////////
-void do_nothing(pika::lcos::barrier& b1, pika::lcos::local::barrier& b2)
+void do_nothing(pika::barrier<>& b1, pika::barrier<>& b2)
 {
-    b1.wait();
+    b1.arrive_and_wait();
     std::this_thread::sleep_for(
         std::chrono::milliseconds(100));    // wait for 100 ms
-    b2.wait();
+    b2.arrive_and_wait();
 }
 
 void test_thread_id_for_default_constructed_thread_is_default_constructed_id()
@@ -35,42 +35,42 @@ void test_thread_id_for_default_constructed_thread_is_default_constructed_id()
 
 void test_thread_id_for_running_thread_is_not_default_constructed_id()
 {
-    pika::lcos::barrier b1(2);
-    pika::lcos::barrier b2(2);
+    pika::barrier<> b1(2);
+    pika::barrier<> b2(2);
     pika::thread t(&do_nothing, std::ref(b1), std::ref(b2));
-    b1.wait();
+    b1.arrive_and_wait();
 
     PIKA_TEST_NEQ(t.get_id(), pika::thread::id());
 
-    b2.wait();
+    b2.arrive_and_wait();
     t.join();
 }
 
 void test_different_threads_have_different_ids()
 {
-    pika::lcos::barrier b1(3);
-    pika::lcos::barrier b2(3);
+    pika::barrier<> b1(3);
+    pika::barrier<> b2(3);
 
     pika::thread t(&do_nothing, std::ref(b1), std::ref(b2));
     pika::thread t2(&do_nothing, std::ref(b1), std::ref(b2));
-    b1.wait();
+    b1.arrive_and_wait();
 
     PIKA_TEST_NEQ(t.get_id(), t2.get_id());
 
-    b2.wait();
+    b2.arrive_and_wait();
     t.join();
     t2.join();
 }
 
 void test_thread_ids_have_a_total_order()
 {
-    pika::lcos::barrier b1(4);
-    pika::lcos::barrier b2(4);
+    pika::barrier<> b1(4);
+    pika::barrier<> b2(4);
 
     pika::thread t1(&do_nothing, std::ref(b1), std::ref(b2));
     pika::thread t2(&do_nothing, std::ref(b1), std::ref(b2));
     pika::thread t3(&do_nothing, std::ref(b1), std::ref(b2));
-    b1.wait();
+    b1.arrive_and_wait();
 
     pika::thread::id t1_id = t1.get_id();
     pika::thread::id t2_id = t2.get_id();
@@ -156,7 +156,7 @@ void test_thread_ids_have_a_total_order()
     PIKA_TEST(!(default_id >= t2_id));
     PIKA_TEST(!(default_id >= t3_id));
 
-    b2.wait();
+    b2.arrive_and_wait();
 
     t1.join();
     t2.join();
