+++ tests/performance/local/htts_v2/htts2_pika.cpp
@@ -117,7 +117,7 @@ private:
         }
     }
 
-    void wait_for_tasks(pika::lcos::barrier& finished)
+    void wait_for_tasks(pika::barrier<>& finished)
     {
         std::uint64_t const pending_count = pika::threads::get_thread_count(
             pika::execution::thread_priority::normal,
@@ -140,7 +140,7 @@ private:
             }
         }
 
-        finished.wait();
+        finished.arrive_and_drop();
     }
 
     using results_type = double;
@@ -194,7 +194,7 @@ private:
         // Schedule a low-priority thread; when it is executed, it checks to
         // make sure all the tasks (which are normal priority) have been
         // executed, and then it
-        pika::lcos::barrier finished(2);
+        pika::barrier<> finished(2);
 
         pika::threads::detail::thread_init_data data(
             pika::threads::detail::make_thread_function_nullary(
@@ -203,7 +203,7 @@ private:
             nullptr, pika::execution::thread_priority::low);
         register_work(data);
 
-        finished.wait();
+        finished.arrive_and_wait();
 
         // w_M [nanoseconds]
         results = static_cast<double>(t.elapsed());
