+++ tests/regressions/threads/thread_suspend_duration.cpp
@@ -23,7 +23,7 @@ using pika::program_options::variables_map;
 
 using pika::threads::detail::register_work;
 
-using pika::lcos::barrier;
+using pika::barrier;
 
 using std::chrono::microseconds;
